:truck: Rename(auth): auth 컴포넌트 위치 수정
